Implementar servicios y controladores para Pesca, Flota y Viaje; agregar validaciones y mejorar manejo de errores

+++ backend/src/pesca/uses-cases/PescaService.ts
@@ -1,37 +1,86 @@
-import { Pesca, IPescaRepository,PescaDTO } from "@/pesca/domain";
+import {
+  IPescaRepository,
+  IFlotaRepository,
+  IViajeRepository,
+  PescaDTO,
+  Pesca,
+} from "@/pesca/domain";
+import { AppError } from "@/shared/errors/AppError";
 
 export class PescaService {
   private readonly repository: IPescaRepository;
-
-  constructor(repository: IPescaRepository) {
+  private readonly flotaRepository: IFlotaRepository;
+  private readonly viajeRepository: IViajeRepository;
+  constructor(
+    repository: IPescaRepository,
+    flotaRepository: IFlotaRepository,
+    viajeRepository: IViajeRepository
+  ) {
     this.repository = repository;
+    this.flotaRepository = flotaRepository;
+    this.viajeRepository = viajeRepository;
   }
 
   async getPescas() {
-    return this.repository.getPescas();
+    return await this.repository.getPescas();
   }
 
-  async getPescaById(id: number) {
-    return this.repository.getPescaById(id);
+  async getPescaById(id: number): Promise<Pesca> {
+    const pesca = await this.repository.getPescaById(id);
+    if (!pesca) {
+      throw new AppError("Pesca no encontrada", 404);
+    }
+    return pesca;
   }
 
   async createPesca(pesca: PescaDTO) {
-    return this.repository.createPesca(pesca);
+    if (!pesca.id_viaje) {
+      throw new AppError("Viaje de pesca requerido", 400);
+    }
+    if (!pesca.pescado_tipo) {
+      throw new AppError("Tipo de pescado requerido", 400);
+    }
+    if (!pesca.pescado_cajas) {
+      throw new AppError("Cajas de pescado requeridas", 400);
+    }
+    const pescaCreated = await this.repository.createPesca(pesca);
+    return pescaCreated;
   }
 
   async updatePesca(id: number, pesca: PescaDTO) {
+    if (!(await this.getPescaById(id))) {
+      throw new AppError("Pesca no encontrada", 404);
+    }
+    if (!pesca.id_viaje && !pesca.pescado_tipo && !pesca.pescado_cajas) {
+      throw new AppError("Datos de pesca requeridos", 400);
+    }
     this.repository.updatePesca(id, pesca);
   }
 
   async deletePesca(id: number) {
+    if (!(await this.getPescaById(id))) {
+      throw new AppError("Pesca no encontrada", 404);
+    }
     this.repository.deletePesca(id);
   }
 
   async getPescaByViajeId(id: number) {
-    return this.repository.getPescaByViajeId(id);
+    if (!id) {
+      throw new AppError("Id de viaje requerido", 400);
+    }
+    if (!(await this.viajeRepository.getViajeById(id))) {
+      throw new AppError("Viaje no encontrado", 404);
+    }
+    return await this.repository.getPescaByViajeId(id);
   }
 
   async getPescaByFlotaId(id: number) {
-    return this.repository.getPescaByFlotaId(id);
+    if (!id) {
+      throw new AppError("Id de flota requerido", 400);
+    }
+    if (!(await this.flotaRepository.getFlotaById(id))) {
+      throw new AppError("Flota no encontrada", 404);
+    }
+    return await this.repository.getPescaByFlotaId(id);
   }
 }
